refactor(cli): extract lifecycle command module (start, serve, stop, reboot, logs, setup)
resolveServerLaunchSpec's project-root resolution accounts for the new
module depth (src/cli/commands/ is three levels below the repo root where
cli/ was one); asserted by legacy-start.test.ts. The src/core/cli
dependency stays dynamic in both directions.
